Refactoring common UC/ESA content into base class
