+++ src/main/java/uk/gov/hmcts/reform/sscs/ccd/presubmit/writefinaldecision/WriteFinalDecisionComponentId.java
@@ -1,0 +1,5 @@
+package uk.gov.hmcts.reform.sscs.ccd.presubmit.writefinaldecision;
+
+public enum WriteFinalDecisionComponentId {
+    REASON, ANYTHING_ELSE, RECOMMENDATION, HEARING_TYPE;
+}
