+++ src/main/java/uk/gov/hmcts/reform/sscs/ccd/presubmit/writefinaldecision/uc/UcNewTemplateContent.java
@@ -1,33 +1,15 @@
 package uk.gov.hmcts.reform.sscs.ccd.presubmit.writefinaldecision.uc;
 
-
-import static java.util.Arrays.asList;
-import static java.util.Collections.singletonList;
-import static org.apache.commons.lang3.StringUtils.equalsIgnoreCase;
-
-import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.util.List;
 import uk.gov.hmcts.reform.sscs.ccd.presubmit.writefinaldecision.uc.scenarios.UcScenario;
-import uk.gov.hmcts.reform.sscs.ccd.presubmit.writefinaldecision.uc.scenarios.UcTemplateComponentId;
 import uk.gov.hmcts.reform.sscs.model.docassembly.Descriptor;
-import uk.gov.hmcts.reform.sscs.model.docassembly.Paragraph;
-import uk.gov.hmcts.reform.sscs.model.docassembly.WriteFinalDecisionTemplateBody;
 
 public abstract class UcNewTemplateContent extends UcTemplateContent {
 
     protected static DateTimeFormatter DATE_FORMATTER = DateTimeFormatter.ofPattern("dd/MM/yyyy");
 
     public abstract UcScenario getScenario();
-
-    public String getAllowedOrRefusedSentence(boolean allowed) {
-        return "The appeal is " + (allowed ? "allowed" : "refused") + ".";
-    }
-
-    public String getConfirmedOrSetAsideSentence(boolean setAside, String decisionDate) {
-        return "The decision made by the Secretary of State on " + DATE_FORMATTER.format(LocalDate.parse(decisionDate)) + " is "
-            + (!setAside ? "confirmed." : "set aside.");
-    }
 
     public String getDoesNotHaveLimitedCapabilityForWorkSentence(String appellantName) {
         return appellantName + " does not have limited capability for work. The matter is now remitted to the Secretary of State to make a final decision upon entitlement to UC.";
@@ -42,44 +24,6 @@ public abstract class UcNewTemplateContent extends UcTemplateContent {
             return "The following activity and descriptor from Schedule 7 of the UC Regulations 2013 applied: ";
         } else {
             return "The following activities and descriptors from Schedule 7 of the UC Regulations 2013 applied: ";
-        }
-    }
-
-    public List<String> getHearingTypeSentences(String appellantName, String bundlePage, String hearingType, boolean appellantAttended, boolean presentingOfficerAttened) {
-        if (equalsIgnoreCase("paper", hearingType)) {
-            return asList("No party has objected to the matter being decided without a hearing.", "Having considered the appeal bundle to page " + bundlePage + " and the requirements of rules 2 and 27 of the Tribunal Procedure (First-tier Tribunal) (Social Entitlement Chamber) Rules 2008 the Tribunal is satisfied that it is able to decide the case in this way.");
-        } else  {
-            return getFaceToFaceTelephoneVideoHearingTypeSentences(hearingType, appellantName, bundlePage, appellantAttended, presentingOfficerAttened);
-        }
-    }
-
-    public List<String> getFaceToFaceTelephoneVideoHearingTypeSentences(String hearingType, String appellantName, String bundlePage,
-        boolean appellantAttended, boolean presentingOfifficerAttened) {
-        if (appellantAttended) {
-            if (equalsIgnoreCase("faceToFace", hearingType)) {
-                return singletonList("This has been an oral (face to face) hearing. "
-                        + getAppellantAttended(hearingType, appellantName, presentingOfifficerAttened, bundlePage));
-            } else {
-                return singletonList("This has been a remote hearing in the form of a " + hearingType + " hearing. "
-                    + getAppellantAttended(hearingType, appellantName, presentingOfifficerAttened, bundlePage));
-            }
-        } else {
-            if (equalsIgnoreCase("faceToFace", hearingType)) {
-                return asList(appellantName + " requested an oral hearing but did not attend today. "
-                    + (presentingOfifficerAttened ? "A " : "No ") + "Presenting Officer attended on behalf of the Respondent.",
-                    getConsideredParagraph(bundlePage, appellantName));
-            } else {
-                return asList("This has been a remote hearing in the form of a " + hearingType + " hearing. " + appellantName + " did not attend the hearing today. "
-                    + (presentingOfifficerAttened ? "A" : "No") + " Presenting Officer attended on behalf of the Respondent.",
-                    getConsideredParagraph(bundlePage, appellantName));
-            }
-        }
-    }
-
-    public void addHearingType(WriteFinalDecisionTemplateBody writeFinalDecisionTemplateBody) {
-        for (String hearingTypeSentence : getHearingTypeSentences(writeFinalDecisionTemplateBody.getAppellantName(), writeFinalDecisionTemplateBody.getPageNumber(),
-            writeFinalDecisionTemplateBody.getHearingType(), writeFinalDecisionTemplateBody.isAttendedHearing(), writeFinalDecisionTemplateBody.isPresentingOfficerAttended())) {
-            addComponent(new Paragraph(UcTemplateComponentId.HEARING_TYPE.name(), hearingTypeSentence));
         }
     }
